refactor: BusinessException 생성, Pagination 모듈 생성

+++ backend/src/main/java/com/bookhub/bookhub_back/service/impl/AuthorServiceImpl.java
@@ -3,16 +3,20 @@ package com.bookhub.bookhub_back.service.impl;
 import com.bookhub.bookhub_back.common.constants.ResponseCode;
 import com.bookhub.bookhub_back.common.constants.ResponseMessage;
 
+import com.bookhub.bookhub_back.common.constants.ResponseMessageKorean;
 import com.bookhub.bookhub_back.dto.ResponseDto;
 import com.bookhub.bookhub_back.dto.author.request.AuthorRequestDto;
 import com.bookhub.bookhub_back.dto.author.response.AuthorResponseDto;
 import com.bookhub.bookhub_back.entity.Author;
+import com.bookhub.bookhub_back.exception.BusinessException;
 import com.bookhub.bookhub_back.exception.DuplicateResourceException;
 import com.bookhub.bookhub_back.repository.AuthorRepository;
+import com.bookhub.bookhub_back.repository.BookRepository;
 import com.bookhub.bookhub_back.service.AuthorService;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -21,11 +25,12 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class AuthorServiceImpl implements AuthorService {
     private final AuthorRepository authorRepository;
+    private final BookRepository bookRepository;
 
     @Override
     public ResponseDto<AuthorResponseDto> createAuthor(AuthorRequestDto dto) {
         if (authorRepository.existsByAuthorEmail(dto.getAuthorEmail())) {
-            throw new DuplicateResourceException("이미 존재하는 작가 이메일입니다.");
+            throw new BusinessException(ResponseCode.DUPLICATED_ENTITY, "이미 존재하는 작가 이메일입니다.");
         }
 
         Author newAuthor = Author.builder()
@@ -42,6 +47,9 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public ResponseDto<List<AuthorResponseDto>> getAuthorsByName(String authorName) {
+        if (authorName == null || authorName.trim().isEmpty()) {
+            throw new IllegalArgumentException(ResponseMessageKorean.INVALID_SEARCH_CONDITION);
+        }
         List<Author> authors = authorRepository.findAllByAuthorNameContaining(authorName);
 
         List<AuthorResponseDto> responseDtos = authors.stream()
@@ -52,13 +60,14 @@ public class AuthorServiceImpl implements AuthorService {
     }
 
     @Override
+    @Transactional
     public ResponseDto<AuthorResponseDto> updateAuthor(Long authorId, AuthorRequestDto dto) {
         Author author = authorRepository.findById(authorId)
-                .orElseThrow(() -> new EntityNotFoundException());
+                .orElseThrow(EntityNotFoundException::new);
 
         if (!author.getAuthorEmail().equals(dto.getAuthorEmail())
                 && authorRepository.existsByAuthorEmail(dto.getAuthorEmail())) {
-            throw new DuplicateResourceException("이미 존재하는 작가 이메일입니다.");
+            throw new BusinessException(ResponseCode.DUPLICATED_ENTITY, "이미 존재하는 작가 이메일입니다.");
         }
 
         author.setAuthorName(dto.getAuthorName());
@@ -72,9 +81,14 @@ public class AuthorServiceImpl implements AuthorService {
     }
 
     @Override
+    @Transactional
     public ResponseDto<Void> deleteAuthor(Long authorId) {
         Author author = authorRepository.findById(authorId)
-                .orElseThrow(() -> new EntityNotFoundException());
+                .orElseThrow(EntityNotFoundException::new);
+
+        if (bookRepository.existsByAuthorId_AuthorId(author.getAuthorId())) {
+            throw new BusinessException(ResponseCode.ENTITY_REFERENCE, "참조 중인 작가는 삭제할 수 없습니다.");
+        }
 
         authorRepository.delete(author);
